+++ client/jetbrains/src/main/java/com/sourcegraph/cody/recipes/ExplainCodeDetailedAction.java
@@ -1,0 +1,14 @@
+package com.sourcegraph.cody.recipes;
+
+public class ExplainCodeDetailedAction extends BaseRecipeAction {
+
+  @Override
+  protected PromptProvider getPromptProvider() {
+    return new ExplainCodeDetailedPromptProvider();
+  }
+
+  @Override
+  protected String getActionComponentName() {
+    return "recipe:explain-code-detailed";
+  }
+}