+++ client/jetbrains/src/main/java/com/sourcegraph/cody/recipes/GenerateDocStringAction.java
@@ -1,0 +1,14 @@
+package com.sourcegraph.cody.recipes;
+
+public class GenerateDocStringAction extends BaseRecipeAction {
+
+  @Override
+  protected PromptProvider getPromptProvider() {
+    return new GenerateDocStringPromptProvider();
+  }
+
+  @Override
+  protected String getActionComponentName() {
+    return "recipe:generate-docstring";
+  }
+}