+++ client/jetbrains/src/main/java/com/sourcegraph/cody/recipes/GenerateUnitTestAction.java
@@ -1,0 +1,14 @@
+package com.sourcegraph.cody.recipes;
+
+public class GenerateUnitTestAction extends BaseRecipeAction {
+
+  @Override
+  protected PromptProvider getPromptProvider() {
+    return new GenerateUnitTestPromptProvider();
+  }
+
+  @Override
+  protected String getActionComponentName() {
+    return "recipe:generate-unit-test";
+  }
+}